pyproject.toml
[project]
name = "openwrite"
version = "5.4.0"
description = "OpenWrite 长篇小说创作引擎 — AI 辅助小说创作"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.18",
    "prompt_toolkit>=3.0",
]

[tool.setuptools.packages.find]
include = ["tools*", "models*"]
exclude = ["tests*", "data*", "craft*", "skills*", "config*", "docs*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
]

[project.scripts]
openwrite = "tools.cli:main"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "reference/",
    "__pycache__/",
    "data/",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
